Fixed merge conflicts and some lint warnings
