+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,27 +1,27 @@
 /*The dependencies are organized in alphabetical order so they are easy to quickly locate,
 please add accordingly*/
 object Dependencies {
-    const val appCompatVersion = "1.4.1"
     const val androidPluginVersion = "7.2.1"
+    const val appCompatVersion = "1.4.1"
     const val cameraxVersion = "1.2.0-alpha04"
     const val composeActVersion = "1.5.1"
     const val composeConstraintVersion = "1.0.1"
     const val composeFoundationVersion = "1.3.0-alpha02"
-    const val composeVersion = "1.2.0"
     const val composeNavigationVersion = "2.5.1"
+    const val composeVersion = "1.2.0"
     const val composeVmVersion = "2.5.1"
     const val constraintLayoutVersion = "2.1.3"
     const val coreVersion = "1.7.0"
     const val firebaseBomVersion = "28.0.0"
-    const val firebaseCrashlyticsVersion = "18.2.10"
     const val firebaseAnalyticsVersion = "21.0.0"
     const val firebaseCrashlyticsGradleVersion = "2.6.1"
+    const val firebaseCrashlyticsVersion = "18.2.10"
     const val firebaseMessagingVersion = "23.0.8"
     const val googleServicesVersion = "4.3.10"
+    const val jacocoVersion = "0.8.6"
     const val koinVersion = "3.2.0"
     const val kotlinVersion = "1.7.0"
     const val ktlintVersion = "10.3.0"
-    const val jacocoVersion = "0.8.6"
     const val leakCanaryVersion = "2.7"
     const val lifecycleRuntimeVersion = "2.4.1"
     const val lifecycleVersion = "2.0.0"
@@ -33,6 +33,7 @@ object Dependencies {
     const val playServicesVersion = "18.0.1"
     const val recyclerViewVersion = "1.2.1"
     const val retrofitVersion = "2.9.0"
+    const val roomVersion = "2.4.3"
     const val rxAndroidVersion = "2.1.1"
     const val rxJavaVersion = "2.2.21"
     const val splashVersion = "1.0.0"
